requirements.txt
asgiref==3.3.1
dependency-injector==4.10.0
Django==3.1.5
django-rest-framework==0.1.0
djangorestframework==3.12.2
mysqlclient==2.0.3
Pillow==8.1.0
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
